+++ testing/conftest.py
@@ -1,5 +1,5 @@
 import pytest
-import pytest_twisted
+import pytest_twisted.core
 
 
 pytest_plugins = "pytester"
@@ -7,4 +7,4 @@ pytest_plugins = "pytester"
 
 @pytest.hookimpl(tryfirst=True)
 def pytest_configure(config):
-    pytest_twisted._use_asyncio_selector_if_required(config=config)
+    pytest_twisted.core._use_asyncio_selector_if_required(config=config)
